requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-hashes
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudevents==1.12.0
colorama==0.4.6 ; sys_platform == 'win32'
dataclass-wizard==0.35.1
deprecation==2.1.0
docstring-parser==0.17.0
flask==3.1.2
functions-framework==3.10.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-aiplatform==1.129.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-genai==1.52.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
slack-sdk==3.39.0
starlette==0.50.0
tenacity==9.1.2
todoist-api-python==3.1.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvicorn-worker==0.4.0
watchdog==6.0.0
websockets==15.0.1
werkzeug==3.1.4
